Compiled Isorropia with -Wall, and fixed a few minor warnings that showed up.
Improved the doxygen commenting for a few functions/methods.

Fixed a couple of bugs that I hadn't noticed. The test-harness was
reporting that the tests passed, but some were failing. I thought
the test harness was just checking the test output for
the string "passed", but apparently that's not the case.
Anyway, the tests are actually passing now. I'll find out how
the test-harness actually detects test results.

+++ packages/isorropia/src/zoltan/Isorropia_ZoltanQuery.cpp
@@ -126,7 +126,7 @@ int Isorropia::ZoltanQuery::Number_Edges  ( void * data,
 {
   int LocalRow = graph_.LRID( *global_id );
 
-  if( LocalRow != -1 && LocalRow == *local_id )
+  if( LocalRow != -1 && LocalRow == (int)*local_id )
   {
     *ierr = ZOLTAN_OK;
